sdcardfs: Handle rm/rename across trees
When renaming or removing, other views must be updated as well. We
do this by looking up the inode corresponding to the lower inodes
in the other sdcardfs mounts.

There are three cases to consider:
1. The other view does not know about that lower inode
2. The other view does not know about the location you're
   moving to.
3. The other view knows about that lower inode and the
   target location.

For case 1, we can just ignore it and move on.
For case 3, we do some clean up work to ensure that the
permissions are handled.
For case 2, we create a fake inode to hold the permissions for
the missing parent.

Since we can't safely take the appropriate locks to fully fix
the parents from this context, we settle for fixing the permissions
If the parent directory of a program's cwd is moved from a different
view, permissions will be respected, but attempting to navigate
relative directories will fail.

Change-Id: Iab5d61e49654ad2b759b825028d538aa303ae909